refactor: moved out helper DMS classes to their own module
