+++ cognite/neat/rules/models/dms/_dms_architect_rules.py
@@ -10,15 +10,8 @@ from typing import TYPE_CHECKING, Any, ClassVar, Literal, cast
 
 from cognite.client import data_modeling as dm
 from cognite.client.data_classes.data_modeling import PropertyType as CognitePropertyType
-from cognite.client.data_classes.data_modeling.containers import BTreeIndex
-from cognite.client.data_classes.data_modeling.views import (
-    SingleEdgeConnectionApply,
-    SingleReverseDirectRelationApply,
-    ViewPropertyApply,
-)
 from pydantic import Field, field_serializer, field_validator, model_serializer, model_validator
 from pydantic_core.core_schema import SerializationInfo, ValidationInfo
-from rdflib import Namespace
 
 import cognite.neat.rules.issues.spreadsheet
 from cognite.neat.rules import issues
@@ -44,19 +37,16 @@ from cognite.neat.rules.models.entities import (
     ClassEntity,
     ContainerEntity,
     ContainerEntityList,
-    DMSNodeEntity,
     DMSUnknownEntity,
-    ParentClassEntity,
     ReferenceEntity,
-    UnknownEntity,
     URLEntity,
     ViewEntity,
     ViewEntityList,
     ViewPropertyEntity,
 )
-from cognite.neat.rules.models.wrapped_entities import DMSFilter, HasDataFilter, NodeTypeFilter
+from cognite.neat.rules.models.wrapped_entities import HasDataFilter, NodeTypeFilter
 
-from ._dms_schema import DMSSchema, PipelineSchema
+from ._dms_schema import DMSSchema
 
 if TYPE_CHECKING:
     from cognite.neat.rules.models.information._information_rules import InformationRules
@@ -609,6 +599,8 @@ class DMSRules(BaseRules):
         handler: Callable,
         info: SerializationInfo,
     ) -> dict[str, Any]:
+        from ._serializer import _DMSRulesSerializer
+
         dumped = cast(dict[str, Any], handler(self, info))
         space, version = self.metadata.space, self.metadata.version
         return _DMSRulesSerializer(info, space, version).clean(dumped)
@@ -616,12 +608,18 @@ class DMSRules(BaseRules):
     def as_schema(
         self, include_ref: bool = False, include_pipeline: bool = False, instance_space: str | None = None
     ) -> DMSSchema:
+        from ._exporter import _DMSExporter
+
         return _DMSExporter(self, include_ref, include_pipeline, instance_space).to_schema()
 
     def as_information_architect_rules(self) -> "InformationRules":
+        from ._converter import _DMSRulesConverter
+
         return _DMSRulesConverter(self).as_information_architect_rules()
 
     def as_domain_expert_rules(self) -> DomainRules:
+        from ._converter import _DMSRulesConverter
+
         return _DMSRulesConverter(self).as_domain_rules()
 
     def reference_self(self) -> Self:
@@ -639,623 +637,3 @@ class DMSRules(BaseRules):
         for container in new_rules.containers or []:
             container.reference = ReferenceEntity(prefix=container.container.prefix, suffix=container.container.suffix)
         return new_rules
-
-
-class _DMSExporter:
-    """The DMS Exporter is responsible for exporting the DMSRules to a DMSSchema.
-
-    This kept in this location such that it can be used by the DMSRules to validate the schema.
-    (This module cannot have a dependency on the exporter module, as it would create a circular dependency.)
-
-    Args
-        include_pipeline (bool): If True, the pipeline will be included with the schema. Pipeline means the
-            raw tables and transformations necessary to populate the data model.
-        instance_space (str): The space to use for the instance. Defaults to None,`Rules.metadata.space` will be used
-    """
-
-    def __init__(
-        self,
-        rules: DMSRules,
-        include_ref: bool = True,
-        include_pipeline: bool = False,
-        instance_space: str | None = None,
-    ):
-        self.include_ref = include_ref
-        self.include_pipeline = include_pipeline
-        self.instance_space = instance_space
-        self.rules = rules
-        self._ref_schema = rules.reference.as_schema() if rules.reference else None
-        if self._ref_schema:
-            # We skip version as that will always be missing in the reference
-            self._ref_views_by_id = {dm.ViewId(view.space, view.external_id): view for view in self._ref_schema.views}
-        else:
-            self._ref_views_by_id = {}
-
-    def to_schema(self) -> DMSSchema:
-        rules = self.rules
-        container_properties_by_id, view_properties_by_id = self._gather_properties()
-        containers = self._create_containers(container_properties_by_id)
-
-        views, node_types = self._create_views_with_node_types(view_properties_by_id)
-
-        views_not_in_model = {view.view.as_id() for view in rules.views if not view.in_model}
-        data_model = rules.metadata.as_data_model()
-        data_model.views = sorted(
-            [view_id for view_id in views.as_ids() if view_id not in views_not_in_model],
-            key=lambda v: v.as_tuple(),  # type: ignore[union-attr]
-        )
-
-        spaces = self._create_spaces(rules.metadata, containers, views, data_model)
-
-        output = DMSSchema(
-            spaces=spaces,
-            data_models=dm.DataModelApplyList([data_model]),
-            views=views,
-            containers=containers,
-            node_types=node_types,
-        )
-        if self.include_pipeline:
-            return PipelineSchema.from_dms(output, self.instance_space)
-
-        if self._ref_schema and self.include_ref:
-            output.frozen_ids.update(self._ref_schema.node_types.as_ids())
-            output.frozen_ids.update(self._ref_schema.views.as_ids())
-            output.frozen_ids.update(self._ref_schema.containers.as_ids())
-            output.node_types.extend(self._ref_schema.node_types)
-            output.views.extend(self._ref_schema.views)
-            output.containers.extend(self._ref_schema.containers)
-            output.data_models.extend(self._ref_schema.data_models)
-
-        return output
-
-    def _create_spaces(
-        self,
-        metadata: DMSMetadata,
-        containers: dm.ContainerApplyList,
-        views: dm.ViewApplyList,
-        data_model: dm.DataModelApply,
-    ) -> dm.SpaceApplyList:
-        used_spaces = {container.space for container in containers} | {view.space for view in views}
-        if len(used_spaces) == 1:
-            # We skip the default space and only use this space for the data model
-            data_model.space = used_spaces.pop()
-            spaces = dm.SpaceApplyList([dm.SpaceApply(space=data_model.space)])
-        else:
-            used_spaces.add(metadata.space)
-            spaces = dm.SpaceApplyList([dm.SpaceApply(space=space) for space in used_spaces])
-        if self.instance_space and self.instance_space not in {space.space for space in spaces}:
-            spaces.append(dm.SpaceApply(space=self.instance_space, name=self.instance_space))
-        return spaces
-
-    def _create_views_with_node_types(
-        self,
-        view_properties_by_id: dict[dm.ViewId, list[DMSProperty]],
-    ) -> tuple[dm.ViewApplyList, dm.NodeApplyList]:
-        views = dm.ViewApplyList([dms_view.as_view() for dms_view in self.rules.views])
-        dms_view_by_id = {dms_view.view.as_id(): dms_view for dms_view in self.rules.views}
-
-        for view in views:
-            view_id = view.as_id()
-            view.properties = {}
-            if not (view_properties := view_properties_by_id.get(view_id)):
-                continue
-            for prop in view_properties:
-                view_property = self._create_view_property(prop, view_properties_by_id)
-                if view_property is not None:
-                    view.properties[prop.view_property] = view_property
-
-        data_model_type = self.rules.metadata.data_model_type
-        unique_node_types: set[dm.NodeId] = set()
-        parent_views = {parent for view in views for parent in view.implements or []}
-        for view in views:
-            dms_view = dms_view_by_id.get(view.as_id())
-            dms_properties = view_properties_by_id.get(view.as_id(), [])
-            view_filter = self._create_view_filter(view, dms_view, data_model_type, dms_properties)
-
-            view.filter = view_filter.as_dms_filter()
-
-            if isinstance(view_filter, NodeTypeFilter):
-                unique_node_types.update(view_filter.nodes)
-                if view.as_id() in parent_views:
-                    warnings.warn(issues.dms.NodeTypeFilterOnParentViewWarning(view.as_id()), stacklevel=2)
-            elif isinstance(view_filter, HasDataFilter) and data_model_type == DataModelType.solution:
-                if dms_view and isinstance(dms_view.reference, ReferenceEntity):
-                    references = {dms_view.reference.as_view_id()}
-                elif any(True for prop in dms_properties if isinstance(prop.reference, ReferenceEntity)):
-                    references = {
-                        prop.reference.as_view_id()
-                        for prop in dms_properties
-                        if isinstance(prop.reference, ReferenceEntity)
-                    }
-                else:
-                    continue
-                warnings.warn(
-                    issues.dms.HasDataFilterOnViewWithReferencesWarning(view.as_id(), list(references)), stacklevel=2
-                )
-
-        return views, dm.NodeApplyList(
-            [dm.NodeApply(space=node.space, external_id=node.external_id) for node in unique_node_types]
-        )
-
-    @classmethod
-    def _create_edge_type_from_prop(cls, prop: DMSProperty) -> dm.DirectRelationReference:
-        if isinstance(prop.reference, ReferenceEntity):
-            ref_view_prop = prop.reference.as_view_property_id()
-            return cls._create_edge_type_from_view_id(cast(dm.ViewId, ref_view_prop.source), ref_view_prop.property)
-        else:
-            return cls._create_edge_type_from_view_id(prop.view.as_id(), prop.view_property)
-
-    @staticmethod
-    def _create_edge_type_from_view_id(view_id: dm.ViewId, property_: str) -> dm.DirectRelationReference:
-        return dm.DirectRelationReference(
-            space=view_id.space,
-            # This is the same convention as used when converting GraphQL to DMS
-            external_id=f"{view_id.external_id}.{property_}",
-        )
-
-    def _create_containers(
-        self,
-        container_properties_by_id: dict[dm.ContainerId, list[DMSProperty]],
-    ) -> dm.ContainerApplyList:
-        containers = dm.ContainerApplyList(
-            [dms_container.as_container() for dms_container in self.rules.containers or []]
-        )
-        container_to_drop = set()
-        for container in containers:
-            container_id = container.as_id()
-            if not (container_properties := container_properties_by_id.get(container_id)):
-                warnings.warn(issues.dms.EmptyContainerWarning(container_id=container_id), stacklevel=2)
-                container_to_drop.add(container_id)
-                continue
-            for prop in container_properties:
-                if prop.container_property is None:
-                    continue
-                if isinstance(prop.value_type, DataType):
-                    type_cls = prop.value_type.dms
-                else:
-                    type_cls = dm.DirectRelation
-
-                type_ = type_cls(is_list=prop.is_list or False)
-                container.properties[prop.container_property] = dm.ContainerProperty(
-                    type=type_,
-                    nullable=prop.nullable if prop.nullable is not None else True,
-                    default_value=prop.default,
-                    name=prop.name,
-                    description=prop.description,
-                )
-
-            uniqueness_properties: dict[str, set[str]] = defaultdict(set)
-            for prop in container_properties:
-                if prop.container_property is not None:
-                    for constraint in prop.constraint or []:
-                        uniqueness_properties[constraint].add(prop.container_property)
-            for constraint_name, properties in uniqueness_properties.items():
-                container.constraints = container.constraints or {}
-                container.constraints[constraint_name] = dm.UniquenessConstraint(properties=list(properties))
-
-            index_properties: dict[str, set[str]] = defaultdict(set)
-            for prop in container_properties:
-                if prop.container_property is not None:
-                    for index in prop.index or []:
-                        index_properties[index].add(prop.container_property)
-            for index_name, properties in index_properties.items():
-                container.indexes = container.indexes or {}
-                container.indexes[index_name] = BTreeIndex(properties=list(properties))
-
-        # We might drop containers we convert direct relations of list into multi-edge connections
-        # which do not have a container.
-        for container in containers:
-            if container.constraints:
-                container.constraints = {
-                    name: const
-                    for name, const in container.constraints.items()
-                    if not (isinstance(const, dm.RequiresConstraint) and const.require in container_to_drop)
-                }
-        return dm.ContainerApplyList(
-            [container for container in containers if container.as_id() not in container_to_drop]
-        )
-
-    def _gather_properties(self) -> tuple[dict[dm.ContainerId, list[DMSProperty]], dict[dm.ViewId, list[DMSProperty]]]:
-        container_properties_by_id: dict[dm.ContainerId, list[DMSProperty]] = defaultdict(list)
-        view_properties_by_id: dict[dm.ViewId, list[DMSProperty]] = defaultdict(list)
-        for prop in self.rules.properties:
-            view_id = prop.view.as_id()
-            view_properties_by_id[view_id].append(prop)
-
-            if prop.container and prop.container_property:
-                container_id = prop.container.as_id()
-                container_properties_by_id[container_id].append(prop)
-
-        return container_properties_by_id, view_properties_by_id
-
-    def _create_view_filter(
-        self,
-        view: dm.ViewApply,
-        dms_view: DMSView | None,
-        data_model_type: DataModelType,
-        dms_properties: list[DMSProperty],
-    ) -> DMSFilter:
-        selected_filter_name = (dms_view and dms_view.filter_ and dms_view.filter_.name) or ""
-        if dms_view and dms_view.filter_ and not dms_view.filter_.is_empty:
-            # Has Explicit Filter, use it
-            return dms_view.filter_
-
-        if data_model_type == DataModelType.solution and selected_filter_name in [NodeTypeFilter.name, ""]:
-            if (
-                dms_view
-                and isinstance(dms_view.reference, ReferenceEntity)
-                and not dms_properties
-                and (ref_view := self._ref_views_by_id.get(dms_view.reference.as_view_id()))
-                and ref_view.filter
-            ):
-                # No new properties, only reference, reuse the reference filter
-                return DMSFilter.from_dms_filter(ref_view.filter)
-            else:
-                referenced_node_ids = {
-                    prop.reference.as_node_entity()
-                    for prop in dms_properties
-                    if isinstance(prop.reference, ReferenceEntity)
-                }
-                if dms_view and isinstance(dms_view.reference, ReferenceEntity):
-                    referenced_node_ids.add(dms_view.reference.as_node_entity())
-                if referenced_node_ids:
-                    return NodeTypeFilter(inner=list(referenced_node_ids))
-
-        # Enterprise Model or (Solution + HasData)
-        ref_containers = view.referenced_containers()
-        if not ref_containers or selected_filter_name == HasDataFilter.name:
-            # Child filter without container properties
-            if selected_filter_name == HasDataFilter.name:
-                warnings.warn(issues.dms.HasDataFilterOnNoPropertiesViewWarning(view.as_id()), stacklevel=2)
-            return NodeTypeFilter(inner=[DMSNodeEntity(space=view.space, externalId=view.external_id)])
-        else:
-            # HasData or not provided (this is the default)
-            return HasDataFilter(inner=[ContainerEntity.from_id(id_) for id_ in ref_containers])
-
-    def _create_view_property(
-        self, prop: DMSProperty, view_properties_by_id: dict[dm.ViewId, list[DMSProperty]]
-    ) -> ViewPropertyApply | None:
-        if prop.container and prop.container_property:
-            container_prop_identifier = prop.container_property
-            extra_args: dict[str, Any] = {}
-            if prop.connection == "direct":
-                if isinstance(prop.value_type, ViewEntity):
-                    extra_args["source"] = prop.value_type.as_id()
-                elif isinstance(prop.value_type, DMSUnknownEntity):
-                    extra_args["source"] = None
-                else:
-                    # Should have been validated.
-                    raise ValueError(
-                        "If this error occurs it is a bug in NEAT, please report"
-                        f"Debug Info, Invalid valueType direct: {prop.model_dump_json()}"
-                    )
-            elif prop.connection is not None:
-                # Should have been validated.
-                raise ValueError(
-                    "If this error occurs it is a bug in NEAT, please report"
-                    f"Debug Info, Invalid connection: {prop.model_dump_json()}"
-                )
-            return dm.MappedPropertyApply(
-                container=prop.container.as_id(),
-                container_property_identifier=container_prop_identifier,
-                name=prop.name,
-                description=prop.description,
-                **extra_args,
-            )
-        elif prop.connection == "edge":
-            if isinstance(prop.value_type, ViewEntity):
-                source_view_id = prop.value_type.as_id()
-            else:
-                # Should have been validated.
-                raise ValueError(
-                    "If this error occurs it is a bug in NEAT, please report"
-                    f"Debug Info, Invalid valueType edge: {prop.model_dump_json()}"
-                )
-            edge_cls: type[dm.EdgeConnectionApply] = dm.MultiEdgeConnectionApply
-            # If is_list is not set, we default to a MultiEdgeConnection
-            if prop.is_list is False:
-                edge_cls = SingleEdgeConnectionApply
-
-            return edge_cls(
-                type=self._create_edge_type_from_prop(prop),
-                source=source_view_id,
-                direction="outwards",
-                name=prop.name,
-                description=prop.description,
-            )
-        elif prop.connection == "reverse":
-            reverse_prop_id: str | None = None
-            if isinstance(prop.value_type, ViewPropertyEntity):
-                source_view_id = prop.value_type.as_view_id()
-                reverse_prop_id = prop.value_type.property_
-            elif isinstance(prop.value_type, ViewEntity):
-                source_view_id = prop.value_type.as_id()
-            else:
-                # Should have been validated.
-                raise ValueError(
-                    "If this error occurs it is a bug in NEAT, please report"
-                    f"Debug Info, Invalid valueType reverse connection: {prop.model_dump_json()}"
-                )
-            reverse_prop: DMSProperty | None = None
-            if reverse_prop_id is not None:
-                reverse_prop = next(
-                    (
-                        prop
-                        for prop in view_properties_by_id.get(source_view_id, [])
-                        if prop.property_ == reverse_prop_id
-                    ),
-                    None,
-                )
-
-            if reverse_prop is None:
-                warnings.warn(
-                    issues.dms.ReverseRelationMissingOtherSideWarning(source_view_id, prop.view_property),
-                    stacklevel=2,
-                )
-
-            if reverse_prop is None or reverse_prop.connection == "edge":
-                inwards_edge_cls = (
-                    dm.MultiEdgeConnectionApply if prop.is_list in [True, None] else SingleEdgeConnectionApply
-                )
-                return inwards_edge_cls(
-                    type=self._create_edge_type_from_prop(reverse_prop or prop),
-                    source=source_view_id,
-                    name=prop.name,
-                    description=prop.description,
-                    direction="inwards",
-                )
-            elif reverse_prop_id and reverse_prop and reverse_prop.connection == "direct":
-                reverse_direct_cls = (
-                    dm.MultiReverseDirectRelationApply if prop.is_list is True else SingleReverseDirectRelationApply
-                )
-                return reverse_direct_cls(
-                    source=source_view_id,
-                    through=dm.PropertyId(source=source_view_id, property=reverse_prop_id),
-                    name=prop.name,
-                    description=prop.description,
-                )
-            else:
-                return None
-
-        elif prop.view and prop.view_property and prop.connection:
-            warnings.warn(
-                issues.dms.UnsupportedConnectionWarning(prop.view.as_id(), prop.view_property, prop.connection or ""),
-                stacklevel=2,
-            )
-        return None
-
-
-class _DMSRulesConverter:
-    def __init__(self, dms: DMSRules):
-        self.dms = dms
-
-    def as_domain_rules(self) -> "DomainRules":
-        raise NotImplementedError("DomainRules not implemented yet")
-
-    def as_information_architect_rules(
-        self,
-        created: datetime | None = None,
-        updated: datetime | None = None,
-        name: str | None = None,
-        namespace: Namespace | None = None,
-    ) -> "InformationRules":
-        from cognite.neat.rules.models.information._information_rules import (
-            InformationClass,
-            InformationMetadata,
-            InformationProperty,
-            InformationRules,
-        )
-
-        dms = self.dms.metadata
-        prefix = dms.space
-
-        metadata = InformationMetadata(
-            schema_=dms.schema_,
-            prefix=prefix,
-            namespace=namespace or Namespace(f"https://purl.orgl/neat/{prefix}/"),
-            version=dms.version,
-            name=name or dms.name or "Missing name",
-            creator=dms.creator,
-            created=dms.created or created or datetime.now(),
-            updated=dms.updated or updated or datetime.now(),
-        )
-
-        classes = [
-            InformationClass(
-                # we do not want a version in class as we use URI for the class
-                class_=ClassEntity(prefix=view.class_.prefix, suffix=view.class_.suffix),
-                description=view.description,
-                parent=[
-                    # we do not want a version in class as we use URI for the class
-                    ParentClassEntity(prefix=implemented_view.prefix, suffix=implemented_view.suffix)
-                    # We only want parents in the same namespace, parent in a different namespace is a reference
-                    for implemented_view in view.implements or []
-                    if implemented_view.prefix == view.class_.prefix
-                ],
-                reference=self._get_class_reference(view),
-            )
-            for view in self.dms.views
-        ]
-
-        properties: list[InformationProperty] = []
-        value_type: DataType | ClassEntity | str
-        for property_ in self.dms.properties:
-            if isinstance(property_.value_type, DataType):
-                value_type = property_.value_type
-            elif isinstance(property_.value_type, ViewEntity | ViewPropertyEntity):
-                value_type = ClassEntity(
-                    prefix=property_.value_type.prefix,
-                    suffix=property_.value_type.suffix,
-                )
-            elif isinstance(property_.value_type, DMSUnknownEntity):
-                value_type = UnknownEntity()
-            else:
-                raise ValueError(f"Unsupported value type: {property_.value_type.type_}")
-
-            properties.append(
-                InformationProperty(
-                    # Removing version
-                    class_=ClassEntity(suffix=property_.class_.suffix, prefix=property_.class_.prefix),
-                    property_=property_.view_property,
-                    value_type=value_type,
-                    description=property_.description,
-                    min_count=0 if property_.nullable or property_.nullable is None else 1,
-                    max_count=float("inf") if property_.is_list or property_.nullable is None else 1,
-                    reference=self._get_property_reference(property_),
-                )
-            )
-
-        return InformationRules(
-            metadata=metadata,
-            properties=SheetList[InformationProperty](data=properties),
-            classes=SheetList[InformationClass](data=classes),
-            reference=self.dms.reference and self.dms.reference.as_information_architect_rules(),  # type: ignore[arg-type]
-        )
-
-    @classmethod
-    def _get_class_reference(cls, view: DMSView) -> ReferenceEntity | None:
-        parents_other_namespace = [parent for parent in view.implements or [] if parent.prefix != view.class_.prefix]
-        if len(parents_other_namespace) == 0:
-            return None
-        if len(parents_other_namespace) > 1:
-            warnings.warn(
-                issues.dms.MultipleReferenceWarning(
-                    view_id=view.view.as_id(), implements=[v.as_id() for v in parents_other_namespace]
-                ),
-                stacklevel=2,
-            )
-        other_parent = parents_other_namespace[0]
-
-        return ReferenceEntity(prefix=other_parent.prefix, suffix=other_parent.suffix)
-
-    @classmethod
-    def _get_property_reference(cls, property_: DMSProperty) -> ReferenceEntity | None:
-        has_container_other_namespace = property_.container and property_.container.prefix != property_.class_.prefix
-        if not has_container_other_namespace:
-            return None
-        container = cast(ContainerEntity, property_.container)
-        return ReferenceEntity(
-            prefix=container.prefix,
-            suffix=container.suffix,
-            property=property_.container_property,
-        )
-
-
-class _DMSRulesSerializer:
-    # These are the fields that need to be cleaned from the default space and version
-    PROPERTIES_FIELDS: ClassVar[list[str]] = ["class_", "view", "value_type", "container"]
-    VIEWS_FIELDS: ClassVar[list[str]] = ["class_", "view", "implements"]
-    CONTAINERS_FIELDS: ClassVar[list[str]] = ["class_", "container"]
-
-    def __init__(self, info: SerializationInfo, default_space: str, default_version: str) -> None:
-        self.default_space = f"{default_space}:"
-        self.default_version = f"version={default_version}"
-        self.default_version_wrapped = f"({self.default_version})"
-
-        self.properties_fields = self.PROPERTIES_FIELDS
-        self.views_fields = self.VIEWS_FIELDS
-        self.containers_fields = self.CONTAINERS_FIELDS
-        self.prop_name = "properties"
-        self.view_name = "views"
-        self.container_name = "containers"
-        self.metadata_name = "metadata"
-        self.prop_view = "view"
-        self.prop_view_property = "view_property"
-        self.prop_value_type = "value_type"
-        self.view_view = "view"
-        self.view_implements = "implements"
-        self.container_container = "container"
-        self.container_constraint = "constraint"
-
-        if info.by_alias:
-            self.properties_fields = [
-                DMSProperty.model_fields[field].alias or field for field in self.properties_fields
-            ]
-            self.views_fields = [DMSView.model_fields[field].alias or field for field in self.views_fields]
-            self.container_fields = [
-                DMSContainer.model_fields[field].alias or field for field in self.containers_fields
-            ]
-            self.prop_view = DMSProperty.model_fields[self.prop_view].alias or self.prop_view
-            self.prop_view_property = DMSProperty.model_fields[self.prop_view_property].alias or self.prop_view_property
-            self.prop_value_type = DMSProperty.model_fields[self.prop_value_type].alias or self.prop_value_type
-            self.view_view = DMSView.model_fields[self.view_view].alias or self.view_view
-            self.view_implements = DMSView.model_fields[self.view_implements].alias or self.view_implements
-            self.container_container = (
-                DMSContainer.model_fields[self.container_container].alias or self.container_container
-            )
-            self.container_constraint = (
-                DMSContainer.model_fields[self.container_constraint].alias or self.container_constraint
-            )
-            self.prop_name = DMSRules.model_fields[self.prop_name].alias or self.prop_name
-            self.view_name = DMSRules.model_fields[self.view_name].alias or self.view_name
-            self.container_name = DMSRules.model_fields[self.container_name].alias or self.container_name
-            self.metadata_name = DMSRules.model_fields[self.metadata_name].alias or self.metadata_name
-
-        if isinstance(info.exclude, dict):
-            # Just for happy mypy
-            exclude = cast(dict, info.exclude)
-            self.exclude_properties = exclude.get("properties", {}).get("__all__", set())
-            self.exclude_views = exclude.get("views", {}).get("__all__", set()) or set()
-            self.exclude_containers = exclude.get("containers", {}).get("__all__", set()) or set()
-            self.metadata_exclude = exclude.get("metadata", set()) or set()
-            self.exclude_top = {k for k, v in exclude.items() if not v}
-        else:
-            self.exclude_top = set(info.exclude or {})
-            self.exclude_properties = set()
-            self.exclude_views = set()
-            self.exclude_containers = set()
-            self.metadata_exclude = set()
-
-    def clean(self, dumped: dict[str, Any]) -> dict[str, Any]:
-        # Sorting to get a deterministic order
-        dumped[self.prop_name] = sorted(
-            dumped[self.prop_name]["data"], key=lambda p: (p[self.prop_view], p[self.prop_view_property])
-        )
-        dumped[self.view_name] = sorted(dumped[self.view_name]["data"], key=lambda v: v[self.view_view])
-        if self.container_name in dumped:
-            dumped[self.container_name] = sorted(
-                dumped[self.container_name]["data"], key=lambda c: c[self.container_container]
-            )
-
-        for prop in dumped[self.prop_name]:
-            for field_name in self.properties_fields:
-                if value := prop.get(field_name):
-                    prop[field_name] = value.removeprefix(self.default_space).removesuffix(self.default_version_wrapped)
-            # Value type can have a property as well
-            prop[self.prop_value_type] = prop[self.prop_value_type].replace(self.default_version, "")
-            if self.exclude_properties:
-                for field in self.exclude_properties:
-                    prop.pop(field, None)
-
-        for view in dumped[self.view_name]:
-            for field_name in self.views_fields:
-                if value := view.get(field_name):
-                    view[field_name] = value.removeprefix(self.default_space).removesuffix(self.default_version_wrapped)
-            if value := view.get(self.view_implements):
-                view[self.view_implements] = ",".join(
-                    parent.strip().removeprefix(self.default_space).removesuffix(self.default_version_wrapped)
-                    for parent in value.split(",")
-                )
-            if self.exclude_views:
-                for field in self.exclude_views:
-                    view.pop(field, None)
-
-        for container in dumped[self.container_name]:
-            for field_name in self.containers_fields:
-                if value := container.get(field_name):
-                    container[field_name] = value.removeprefix(self.default_space)
-
-            if value := container.get(self.container_constraint):
-                container[self.container_constraint] = ",".join(
-                    constraint.strip().removeprefix(self.default_space) for constraint in value.split(",")
-                )
-            if self.exclude_containers:
-                for field in self.exclude_containers:
-                    container.pop(field, None)
-
-        if self.metadata_exclude:
-            for field in self.metadata_exclude:
-                dumped[self.metadata_name].pop(field, None)
-        for field in self.exclude_top:
-            dumped.pop(field, None)
-        return dumped
